+++ src/bit-helpers/index.ts
@@ -1,0 +1,7 @@
+import getComponentsFromJsonAndInline from './getComponentsFromJsonAndInline';
+import getComponentsFromComponentsDir from './getComponentsFromComponentsDir';
+
+export {
+  getComponentsFromJsonAndInline,
+  getComponentsFromComponentsDir,
+};