+++ src/js-completions/index.ts
@@ -1,0 +1,7 @@
+import shouldProvideToJs from './shouldProvideToJs';
+import provideToJs from './provideToJs';
+
+export {
+  shouldProvideToJs,
+  provideToJs,
+};