+++ src/json-completions/index.ts
@@ -1,0 +1,11 @@
+import shouldProvideComponent from './shouldProvideComponent';
+import shouldProvideVersion from './shouldProvideVersion';
+import provideComponent from './provideComponent';
+import provideVersion from './provideVersion';
+
+export {
+  shouldProvideComponent,
+  shouldProvideVersion,
+  provideComponent,
+  provideVersion,
+};